Suxrob

See merge request nurmuhammad.tuhtasinov/ecosystem!218

+++ src/main/java/uz/technocorp/ecosystem/modules/attestation/dto/AttestationReportDto.java
@@ -14,9 +14,9 @@ public class AttestationReportDto {
     private String legalAddress;
     private String hfName;
     private String hfAddress;
-    private Integer totalEmployees;
-    private Integer leadersPassed;
-    private Integer techniciansPassed;
-    private Integer employeesPassed;
-    private Integer failedEmployees;
+    private Long totalEmployees;
+    private Long leadersPassed;
+    private Long techniciansPassed;
+    private Long employeesPassed;
+    private Long failedEmployees;
 }
